crypto: qat - relocate CSR access code
ANBZ: #9185

commit 680302d191b043cf3abe4076794de10171a4ca93 upstream.

Intel-SIG: commit 680302d191b0 crypto: qat - relocate CSR access code
Backport to support Intel QAT live migration for in-tree driver

As the common hw_data files are growing and the adf_hw_csr_ops is going
to be extended with new operations, move all logic related to ring CSRs
to the newly created adf_gen[2|4]_hw_csr_data.[c|h] files.

This does not introduce any functional change.

[ Aichun Shi: amend commit log ]
